+++ src/backends/x11/meta-crtc-xrandr.c
@@ -244,7 +244,7 @@ meta_crtc_xrandr_new (MetaGpuXrandr      *gpu_xrandr,
   all_transforms =
     meta_monitor_transform_from_xrandr_all (xrandr_crtc->rotations);
   crtc_xrandr = g_object_new (META_TYPE_CRTC_XRANDR,
-                              "id", crtc_id,
+                              "id", (uint64_t) crtc_id,
                               "gpu", gpu,
                               "all-transforms", all_transforms,
                               NULL);
