+++ src/backends/x11/meta-output-xrandr.c
@@ -871,7 +871,7 @@ meta_output_xrandr_new (MetaGpuXrandr *gpu_xrandr,
   output_info_init_backlight_limits_xrandr (output_info, xdisplay, output_id);
 
   output = g_object_new (META_TYPE_OUTPUT_XRANDR,
-                         "id", output_id,
+                         "id", (uint64_t) output_id,
                          "gpu", gpu_xrandr,
                          "info", output_info,
                          NULL);
